requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
apache-airflow==2.7.1